move functions to client
